NEVER LOSE a SINGLE MESSAGE by default
after some benchmark testing of ackall vs acklocal:
for a single connection, the throughput will drop from
1500/s to 900/s(40%) in a low latency LAN(0.1ms).

for many concurrent connections, this throughput drop is
not that obvious: still can reach 100K/s on 2 kateways.

so I decide to switch to ackall by default.

another concern after this switch is: it will lower the
availability: if 1 of 2 replica machine is down, Pub
will fail instantly.

anyway, that's worth a try